[www] Rename www/sync.js to www/index.js
- state a clear entry-point to the javascript implementation.
- usually, 'index.js', 'main.js', or the project name are assumed
  as the main entry-point.
- it's not a big deal when there is only one javascript file, but when
  there are multiple files, it's nice to know the main entry-point at a
  glance.
